Ring of Wealth replacement in progress

+++ core/src/main/java/com/shatteredpixel/cursedpixeldungeon/actors/mobs/Mob.java
@@ -671,7 +671,7 @@ public abstract class Mob extends Char {
 		}
 
 		Item luckybadgedrop = LuckyBadge.tryForBonusDrop(Dungeon.hero, 1);
-		if (luckybadgedrop != null) {
+		if (luckybadgedrop != null && Dungeon.hero.belongings.getItem(LuckyBadge.class) != null) {
 			Dungeon.level.drop(luckybadgedrop, pos).sprite.drop();
 		}
